fix(signin): change the order of the login objects
Quote from documentation (*):
"registerForActivityResult() is safe to call
before your fragment or activity is created,
allowing it to be used directly when declaring
member variables for the returned
ActivityResultLauncher instances. Note: While it
is safe to call registerForActivityResult()
before your fragment or activity is created, you
cannot launch the ActivityResultLauncher until
the fragment or activity's Lifecycle has reached
CREATED."

So, move the creation of the login objects (Google
Login and Facebook Login) in the onAttach(context)
method in the Fragment. This allows that
requireActivity() will return always a non-null
activity.

(*) Documentation: https://developer.android.com/training/basics/intents/result#kotlin